Rename black box bounded contexts
