+++ server/src/test/java/io/spine/server/procman/ProcessManagerRepositoryTest.java
@@ -67,7 +67,7 @@ import io.spine.test.procman.event.PmTaskAdded;
 import io.spine.testing.client.TestActorRequestFactory;
 import io.spine.testing.logging.MuteLogging;
 import io.spine.testing.server.ShardingReset;
-import io.spine.testing.server.blackbox.BlackBoxBoundedContext;
+import io.spine.testing.server.blackbox.MultitenantBlackBoxContext;
 import io.spine.testing.server.entity.given.Given;
 import io.spine.type.TypeUrl;
 import org.junit.jupiter.api.AfterEach;
@@ -544,7 +544,7 @@ class ProcessManagerRepositoryTest
                 .newBuilder()
                 .setProjectId(projectId)
                 .build();
-        BlackBoxBoundedContext
+        MultitenantBlackBoxContext
                 .newInstance()
                 .with(new EventDiscardingProcManRepository())
                 .receivesCommand(command)
