+++ testutil-server/src/main/java/io/spine/testing/server/blackbox/EmittedEvents.java
@@ -33,7 +33,7 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- * Provides information on events emitted in the {@link BlackBoxBoundedContext Bounded Context}.
+ * Provides information on events emitted in the {@link MultitenantBlackBoxContext Bounded Context}.
  *
  * @author Mykhailo Drachuk
  */
